break out config support to new source file
